refactor: state names and selection modal isolated
